delete function, add card, update avatar

+++ src/utils/Api.js
@@ -42,6 +42,7 @@ class Api {
       })
       .then(this._checkResponse)
     }
+    
     editProfile({name, about}) {
       return fetch(this._baseUrl + "/users/me", {
         method: "PATCH",
@@ -53,6 +54,7 @@ class Api {
       })
       .then(this._checkResponse)
     }
+    
     removeCard(cardId) {
       return fetch(this._baseUrl + "/cards/" + cardId, {
         method: "DELETE",
@@ -60,21 +62,42 @@ class Api {
       })
       .then(this._checkResponse)
     }
-    //add likes PUT
-    addLikes(cardId){
-      return fetch(this._baseUrl + "/cards/likes/" + cardId, {
-        method: "PUT",
-        headers: this._headers,
-      })
-      .then(this._checkResponse)
+    
+    //add and remove likes
+    changeLikeCardStatus(cardId, like) {
+      if(like) {
+        return fetch(this._baseUrl + "/cards/likes/" + cardId, {
+          method: "PUT",
+          headers: this._headers,
+          body: JSON.stringify({
+            like
+          })
+        })
+        .then(this._checkResponse)
+      } else {
+        return fetch(this._baseUrl + "/cards/likes/" + cardId, {
+          method: "DELETE",
+          headers: this._headers,
+        })
+        .then(this._checkResponse)
+      }
     }
-    removeLike(cardId){
-      return fetch(this._baseUrl + "/cards/likes/" + cardId, {
-        method: "DELETE",
-        headers: this._headers,
-      })
-      .then(this._checkResponse)
-    }
+    // addLikes(cardId){
+    //   return fetch(this._baseUrl + "/cards/likes/" + cardId, {
+    //     method: "PUT",
+    //     headers: this._headers,
+    //   })
+    //   .then(this._checkResponse)
+    // }
+    // removeLike(cardId){
+    //   return fetch(this._baseUrl + "/cards/likes/" + cardId, {
+    //     method: "DELETE",
+    //     headers: this._headers,
+    //   })
+    //   .then(this._checkResponse)
+    // }
+
+
     updateAvatar(avatar) {
       return fetch(this._baseUrl + "/users/me/avatar", {
         method: "PATCH",
